wallet/serializers.py
from rest_framework import serializers
from .models import Wallet, Operation


class WalletSerializer(serializers.ModelSerializer):
    class Meta:
        model = Wallet
        fields = ['uuid', 'balance']


class OperationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Operation
        fields = ['operation_type', 'sum_of_operation']