pyproject.toml
[tool.poetry]
name = "javacode-test"
version = "0.1.0"
description = ""
authors = ["vllscarf <xxxisa2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-rest-framework = "^0.1.0"
psycopg2 = "^2.9.9"
dj-database-url = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
